# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe",
                    "description": "Nome do usuário"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com",
                    "description": "Email do usuário"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456",
                    "description": "Senha do usuário (mínimo 6 caracteres)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1",
                    "name": "John Doe",
                    "email": "john.doe@example.com",
                    "password": "123456",
                    "created_at": "2024-02-23T10:00:00.000Z",
                    "updated_at": "2024-02-23T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "summary": "Cria um usuário",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "507f1f77bcf86cd799439011",
                      "name": "John Doe",
                      "email": "john.doe@example.com",
                      "password": "123456",
                      "created_at": "2024-02-23T10:00:00.000Z",
                      "updated_at": "2024-02-23T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Lista todos os usuários",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "put": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário a ser atualizado",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dados a serem atualizados no usuário",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Updated",
                    "description": "Nome atualizado do usuário"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.updated@example.com",
                    "description": "Email atualizado do usuário"
                  },
                  "password": {
                    "type": "string",
                    "example": "newpassword",
                    "description": "Senha atualizada do usuário (mínimo 6 caracteres)"
                  }
                },
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "507f1f77bcf86cd799439011",
                    "name": "John Updated",
                    "email": "john.updated@example.com",
                    "password": "newpassword",
                    "created_at": "2024-02-23T10:00:00.000Z",
                    "updated_at": "2024-02-23T11:15:30.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "summary": "Atualiza um usuário",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário a ser excluído",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário excluído com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "summary": "Exclui um usuário",
        "tags": [
          "users"
        ]
      }
    },
    "/travel-packages": {
      "post": {
        "operationId": "TravelPackageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Dados do pacote de viagem e imagem (campo \"image\")",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Praia de Maragogi",
                    "description": "Nome do pacote de viagem"
                  },
                  "price": {
                    "type": "number",
                    "example": 1499.99,
                    "description": "Preço da viagem em reais"
                  },
                  "description": {
                    "type": "string",
                    "example": "Uma incrível viagem para as praias paradisíacas de Maragogi...",
                    "description": "Descrição detalhada do pacote de viagem"
                  },
                  "pdfUrl": {
                    "type": "string",
                    "example": "https://example.com/pdf/maragogi-itinerary.pdf",
                    "description": "Link para o PDF com detalhes da viagem"
                  },
                  "maxPeople": {
                    "type": "number",
                    "example": 20,
                    "description": "Número máximo de pessoas para a viagem"
                  },
                  "boardingLocations": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Terminal Tietê - 08:00",
                      "Metrô Tatuapé - 08:30"
                    ],
                    "description": "Locais de embarque"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo da imagem a ser enviado"
                  }
                },
                "required": [
                  "name",
                  "price",
                  "description",
                  "pdfUrl",
                  "maxPeople",
                  "boardingLocations",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pacote de viagem criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPackage"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida. Verifique os dados enviados e se o arquivo de imagem foi informado."
          }
        },
        "tags": [
          "travel-packages"
        ]
      },
      "get": {
        "operationId": "TravelPackageController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de pacotes de viagem retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "1675938274892",
                      "name": "Praia de Maragogi",
                      "price": 1499.99,
                      "description": "Uma incrível viagem para as praias paradisíacas de Maragogi...",
                      "imageUrl": "/travel-packages/1675938274892/image",
                      "pdfUrl": "https://example.com/pdf/maragogi-itinerary.pdf",
                      "maxPeople": 20,
                      "boardingLocations": [
                        "Terminal Tietê - 08:00",
                        "Metrô Tatuapé - 08:30"
                      ],
                      "created_at": "2024-02-23T10:00:00.000Z",
                      "updated_at": "2024-02-23T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Lista todos os pacotes de viagem",
        "tags": [
          "travel-packages"
        ]
      }
    },
    "/travel-packages/{id}": {
      "get": {
        "operationId": "TravelPackageController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do pacote de viagem",
            "schema": {
              "example": "1675938274892",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pacote de viagem encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1675938274892",
                    "name": "Praia de Maragogi",
                    "price": 1499.99,
                    "description": "Uma incrível viagem para as praias paradisíacas de Maragogi...",
                    "imageUrl": "/travel-packages/1675938274892/image",
                    "pdfUrl": "https://example.com/pdf/maragogi-itinerary.pdf",
                    "maxPeople": 20,
                    "boardingLocations": [
                      "Terminal Tietê - 08:00",
                      "Metrô Tatuapé - 08:30"
                    ],
                    "created_at": "2024-02-23T10:00:00.000Z",
                    "updated_at": "2024-02-23T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Pacote de viagem não encontrado"
          }
        },
        "summary": "Busca um pacote de viagem pelo ID",
        "tags": [
          "travel-packages"
        ]
      },
      "put": {
        "operationId": "TravelPackageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do pacote de viagem a ser atualizado",
            "schema": {
              "example": "1675938274892",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Praia de Maragogi"
                  },
                  "price": {
                    "type": "number",
                    "example": 1499.99
                  },
                  "description": {
                    "type": "string",
                    "example": "Uma incrível viagem para as praias paradisíacas de Maragogi..."
                  },
                  "pdfUrl": {
                    "type": "string",
                    "example": "https://example.com/pdf/maragogi-itinerary.pdf"
                  },
                  "maxPeople": {
                    "type": "number",
                    "example": 20
                  },
                  "boardingLocations": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Terminal Tietê - 08:00",
                      "Metrô Tatuapé - 08:30"
                    ]
                  }
                },
                "example": {
                  "name": "Praia de Maragogi",
                  "price": 1499.99,
                  "description": "Uma incrível viagem para as praias paradisíacas de Maragogi...",
                  "pdfUrl": "https://example.com/pdf/maragogi-itinerary.pdf",
                  "maxPeople": 20,
                  "boardingLocations": [
                    "Terminal Tietê - 08:00",
                    "Metrô Tatuapé - 08:30"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pacote de viagem atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "1675938274892",
                    "name": "Praia de Maragogi Atualizado",
                    "price": 1599.99,
                    "description": "Uma incrível viagem para as praias paradisíacas de Maragogi com pacote atualizado...",
                    "imageUrl": "/travel-packages/1675938274892/image",
                    "pdfUrl": "https://example.com/pdf/maragogi-itinerary-updated.pdf",
                    "maxPeople": 25,
                    "boardingLocations": [
                      "Terminal Tietê - 07:30",
                      "Metrô Tatuapé - 08:00",
                      "Shopping Aricanduva - 08:30"
                    ],
                    "created_at": "2024-02-23T10:00:00.000Z",
                    "updated_at": "2024-02-23T11:30:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Pacote de viagem não encontrado"
          }
        },
        "summary": "Atualiza um pacote de viagem",
        "tags": [
          "travel-packages"
        ]
      },
      "delete": {
        "operationId": "TravelPackageController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do pacote de viagem a ser excluído",
            "schema": {
              "example": "1675938274892",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pacote de viagem excluído com sucesso"
          },
          "404": {
            "description": "Pacote de viagem não encontrado"
          }
        },
        "summary": "Exclui um pacote de viagem",
        "tags": [
          "travel-packages"
        ]
      }
    },
    "/travel-packages/{id}/image": {
      "get": {
        "operationId": "TravelPackageController_downloadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem retornada com sucesso",
            "content": {
              "image/jpeg": {}
            }
          },
          "404": {
            "description": "Pacote de viagem ou imagem não encontrada"
          }
        },
        "summary": "Baixar imagem do pacote de viagem",
        "tags": [
          "travel-packages"
        ]
      }
    },
    "/bookings": {
      "post": {
        "operationId": "BookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Dados para criação da reserva",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "507f1f77bcf86cd799439011",
                    "travelPackageId": "507f1f77bcf86cd799439022",
                    "userId": "507f1f77bcf86cd799439033",
                    "fullName": "João Silva",
                    "rg": "12.345.678-9",
                    "cpf": "123.456.789-00",
                    "birthDate": "1990-01-01T00:00:00.000Z",
                    "phone": "(11) 98765-4321",
                    "email": "joao.silva@example.com",
                    "boardingLocation": "Terminal Tietê - 08:00",
                    "created_at": "2024-03-08T10:00:00.000Z",
                    "updated_at": "2024-03-08T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou não há mais vagas disponíveis"
          },
          "404": {
            "description": "Pacote de viagem não encontrado"
          }
        },
        "summary": "Cria uma reserva para um pacote de viagem",
        "tags": [
          "bookings"
        ]
      },
      "get": {
        "operationId": "BookingController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de reservas retornada com sucesso"
          }
        },
        "summary": "Lista todas as reservas",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da reserva",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva encontrada com sucesso"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "summary": "Busca uma reserva pelo ID",
        "tags": [
          "bookings"
        ]
      },
      "delete": {
        "operationId": "BookingController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da reserva a ser cancelada",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reserva cancelada com sucesso"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "summary": "Cancela uma reserva",
        "tags": [
          "bookings"
        ]
      }
    }
  },
  "info": {
    "title": "API do Meu Projeto",
    "description": "Documentação da API do Meu Projeto",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Nome do usuário"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Email do usuário"
          },
          "password": {
            "type": "string",
            "example": "123456",
            "description": "Senha do usuário (mínimo 6 caracteres)"
          }
        }
      },
      "TravelPackage": {
        "type": "object",
        "properties": {}
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "travelPackageId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "ID do pacote de viagem"
          },
          "fullName": {
            "type": "string",
            "example": "João Silva",
            "description": "Nome completo do passageiro"
          },
          "rg": {
            "type": "string",
            "example": "12.345.678-9",
            "description": "RG do passageiro"
          },
          "cpf": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF do passageiro"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-01-01",
            "description": "Data de nascimento do passageiro"
          },
          "phone": {
            "type": "string",
            "example": "(11) 98765-4321",
            "description": "Telefone do passageiro"
          },
          "email": {
            "type": "string",
            "example": "joao.silva@example.com",
            "description": "Email do passageiro"
          },
          "boardingLocation": {
            "type": "string",
            "example": "Terminal Tietê - 08:00",
            "description": "Local de embarque selecionado"
          }
        },
        "required": [
          "travelPackageId",
          "fullName",
          "rg",
          "cpf",
          "birthDate",
          "phone",
          "email",
          "boardingLocation"
        ]
      }
    }
  }
}